Se cambia tipo int a Int64

+++ Modelos/DIRECCION_EMPLEADO.cs
@@ -9,11 +9,11 @@ namespace Modelos
     public class DIRECCION_EMPLEADO
     {
         private int iD;
-        private int dPI;
+        private Int64 dPI;
         private string dIRECCION;
 
         public int ID { get => iD; set => iD = value; }
-        public int DPI { get => dPI; set => dPI = value; }
+        public Int64 DPI { get => dPI; set => dPI = value; }
         public string DIRECCION { get => dIRECCION; set => dIRECCION = value; }
     }
 }
